+++ Actions/Deploy/Deploy.ps1
@@ -19,6 +19,7 @@ $deploymentEnvironments = $deploymentEnvironmentsJson | ConvertFrom-Json | Conve
 $deploymentSettings = $deploymentEnvironments."$environmentName"
 $envName = $environmentName.Split(' ')[0]
 $secrets = $env:Secrets | ConvertFrom-Json
+$settings = $env:Settings | ConvertFrom-Json
 
 # Check obsolete secrets
 "$($envName)-EnvironmentName","$($envName)_EnvironmentName","EnvironmentName" | ForEach-Object {
@@ -34,7 +35,7 @@ $authContext = $null
 foreach($secretName in "$($envName)-AuthContext","$($envName)_AuthContext","AuthContext") {
     if ($secrets."$secretName") {
         $authContext = [System.Text.Encoding]::UTF8.GetString([System.Convert]::FromBase64String($secrets."$secretName"))
-        break
+            break
     }
 }
 if (-not $authContext) {
@@ -110,7 +111,26 @@ else {
 
     try {
         $sandboxEnvironment = ($response.environmentType -eq 1)
-        if ($sandboxEnvironment -and !($bcAuthContext.ClientSecret)) {
+        $scope = $deploymentSettings.Scope
+        if ($null -eq $scope) {
+            if ($settings.Type -eq 'AppSource App' -or ($sandboxEnvironment -and !($bcAuthContext.ClientSecret))) {
+                $scope = 'Dev'
+            }
+            else {
+                $scope = 'PTE'
+            }
+        }
+        elseif (@('Dev','PTE') -notcontains $scope) {
+            throw "Invalid Scope $($scope). Valid values are Dev and PTE."
+        }
+        if (!$sandboxEnvironment -and $type -eq 'CD' -and !($deploymentSettings.continuousDeployment)) {
+            # Continuous deployment is undefined in settings - we will not deploy to production environments
+            Write-Host "::Warning::Ignoring environment $($deploymentSettings.EnvironmentName), which is a production environment"
+        }
+        elseif ($scope -eq 'Dev') {
+            if (!$sandboxEnvironment) {
+                throw "Scope Dev is only valid for sandbox environments"
+            }
             # Sandbox and not S2S -> use dev endpoint (Publish-BcContainerApp)
             $parameters = @{
                 "bcAuthContext" = $bcAuthContext
@@ -126,10 +146,6 @@ else {
             }
             Write-Host "Publishing apps using development endpoint"
             Publish-BcContainerApp @parameters -useDevEndpoint -checkAlreadyInstalled -excludeRuntimePackages -replacePackageId
-        }
-        elseif (!$sandboxEnvironment -and $type -eq 'CD' -and !($deploymentSettings.continuousDeployment)) {
-            # Continuous deployment is undefined in settings - we will not deploy to production environments
-            Write-Host "::Warning::Ignoring environment $($deploymentSettings.EnvironmentName), which is a production environment"
         }
         else {
             # Use automation API for production environments (Publish-PerTenantExtensionApps)