+++ RELEASENOTES.md
@@ -5,6 +5,10 @@ In the summary after a Test Run, you now also have the result of performance tes
 
 ### New Settings
 
+- `deployTo<environmentName>`: is not really new, but has a new property:
+  - **Scope** = specifies the scope of the deployment: Dev, PTE. If not specified, AL-Go for GitHub will always use the Dev Scope for AppSource Apps, but also for PTEs when deploying to sandbox environments when impersonation (refreshtoken) is used for authentication.
+  - **\<custom\>** = custom properties are now supported and will be transferred to a custom deployment script in the hashtable.
+
 - `bcptThresholds` is a JSON object with properties for the default thresholds for the Business Central Performance Toolkit
   - **DurationWarning** - a warning is issued if the duration of a bcpt test degrades more than this percentage (default 10)
   - **DurationError** - an error is issued if the duration of a bcpt test degrades more than this percentage (default 25)